+++ djpcms/tests/test_urls.py
@@ -1,0 +1,9 @@
+from django.test import TestCase
+from django.conf import settings
+
+
+class TestExtraUrls(TestCase):
+    fixtures = ["test.json"]
+    
+    def testExtra(self):
+        pass